refactor package structure for logging
