+++ src/main/java/org/betonquest/betonquest/commands/QuestCommand.java
@@ -28,10 +28,10 @@ import org.betonquest.betonquest.id.EventID;
 import org.betonquest.betonquest.id.ItemID;
 import org.betonquest.betonquest.id.ObjectiveID;
 import org.betonquest.betonquest.item.QuestItem;
+import org.betonquest.betonquest.modules.logger.LogWatcher;
 import org.betonquest.betonquest.utils.PlayerConverter;
 import org.betonquest.betonquest.utils.Utils;
 import org.betonquest.betonquest.utils.location.VectorData;
-import org.betonquest.betonquest.utils.logger.LogWatcher;
 import org.betonquest.betonquest.utils.versioning.Updater;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
